refactor(sdk): split client.test.ts into focused test modules
Split the monolithic test file into separate files mirroring the
source code structure:
- helpers.ts: Shared test utilities (mocks, config)
- text.test.ts: generateText tests
- stream.test.ts: streamText tests
- object.test.ts: generateObject tests
- client.test.ts: KoineError and createKoine factory tests

All 42 tests pass. No functionality changed.
